pre_interview assignment
The shortcut talent team pre-interview assignment contains two programing tasks. 
Task1 is  a function that asks for first name and last name as input and greets the person referring to that person’s full name.
Task2 is  a function that takes your birth year and current year as inputs and outputs your current age. 
The tasks are done using Java programming language.